requirements.txt
docker-py>=0.3.1
pyyaml
jinja2
six
bgtunnel
python-simple-hipchat
